backend/pyproject.toml
[project]
name = "chatbi-backend"
version = "1.0.0"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
    "mysql-connector-python>=8.4.0",
    "sqlalchemy>=2.0.31",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",
    "aiofiles>=23.2.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "numpy>=1.26.4",
    "qwen-agent>=0.0.31",
    "alembic>=1.18.1",
    "redis>=5.0.0",
    "psutil>=7.2.2",
    "pymysql>=1.1.2",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
